refact: WeatherTopAppBar を別ファイルに切り出す
