+++ androidApp/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/ui/weather/compose/WeatherScreen.kt
@@ -1,6 +1,5 @@
 package com.ikmr.banbara23.yaeyama_liner_checker.ui.weather.compose
 
-import androidx.annotation.StringRes
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -12,15 +11,12 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.Scaffold
 import androidx.compose.material.SnackbarDefaults.backgroundColor
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ikmr.banbara23.yaeyama_liner_checker.R
@@ -52,18 +48,6 @@ private fun WeatherScreen(uiState: WeatherUiState, onMoreClick: () -> Unit) {
             uiState = uiState
         )
     }
-}
-
-@Composable
-private fun WeatherTopAppBar(
-    @StringRes titleRes: Int
-) {
-    TopAppBar(
-        contentColor = Color.White,
-        title = { Text(stringResource(titleRes)) },
-        backgroundColor = Color.Transparent,
-        elevation = 0.dp
-    )
 }
 
 @Composable
@@ -102,12 +86,6 @@ private fun WeatherScreenPreview() {
     WeatherScreen(
         WeatherUiState.Loading
     ) {}
-}
-
-@Preview(showBackground = false)
-@Composable
-private fun WeatherTopAppBarPreview() {
-    WeatherTopAppBar(R.string.weather_screen_title)
 }
 
 @Preview(showBackground = true)
